f32 uses fabsf

+++ vlib/builtin/int.v
@@ -30,7 +30,7 @@ pub fn (a f64) eq(b f64) bool {
 	return C.fabs(a - b) <= C.DBL_EPSILON	
 }
 pub fn (a f32) eq(b f32) bool {
-	return C.fabs(a - b) <= C.FLT_EPSILON	
+	return C.fabsf(a - b) <= C.FLT_EPSILON	
 }
 
 // fn (nn i32) str() string {
